+++ service/verify/tests/acceptance/verify_acceptance_test.go
@@ -393,4 +393,5 @@ var _ = Describe("Verify Acceptance Tests", func() {
 
 	// TODO Add Verification Check tests
 	// TODO Add Factor tests
+	// TODO Add Challenge tests
 })