+++ service/verify/v2/service/entity/api_op_client.go
@@ -3,6 +3,8 @@ package entity
 
 import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/entity/challenge"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/entity/challenges"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/entity/factor"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/entity/factors"
 )
@@ -15,8 +17,10 @@ type Client struct {
 	identity   string
 	serviceSid string
 
-	Factor  func(string) *factor.Client
-	Factors *factors.Client
+	Challenge  func(string) *challenge.Client
+	Challenges *challenges.Client
+	Factor     func(string) *factor.Client
+	Factors    *factors.Client
 }
 
 // ClientProperties are the properties required to manage the entity resources
@@ -33,6 +37,17 @@ func New(client *client.Client, properties ClientProperties) *Client {
 		identity:   properties.Identity,
 		serviceSid: properties.ServiceSid,
 
+		Challenge: func(challengeSid string) *challenge.Client {
+			return challenge.New(client, challenge.ClientProperties{
+				Identity:   properties.Identity,
+				ServiceSid: properties.ServiceSid,
+				Sid:        challengeSid,
+			})
+		},
+		Challenges: challenges.New(client, challenges.ClientProperties{
+			Identity:   properties.Identity,
+			ServiceSid: properties.ServiceSid,
+		}),
 		Factor: func(factorSid string) *factor.Client {
 			return factor.New(client, factor.ClientProperties{
 				Identity:   properties.Identity,